% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_music_report.R
\name{clean_music_report}
\alias{clean_music_report}
\title{fix common problems in the Music Report file from EMS}
\usage{
clean_music_report(X)
}
\arguments{
\item{X}{the tibble obtained by reading the Music Report file.}
}
\description{
fix common problems in the Music Report file from EMS
}
