% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_music_files.R
\name{copy_music_files}
\alias{copy_music_files}
\title{Copies music files from orig path to new path}
\usage{
copy_music_files(Tbl)
}
\arguments{
\item{Tbl}{a tibble with at least the columns `orig_music_path` and
`new_music_path`.}
}
\description{
This does not remove all the directories in the destination paths, because
you can do that by hand if you need to.  I don't want you to accidentally
remove a bunch of critical files...
}
