% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_check_curated_table.R
\name{filter_and_check_curated_table}
\alias{filter_and_check_curated_table}
\title{break the curated and edited file into different parts for later use and inspection}
\usage{
filter_and_check_curated_table(path)
}
\arguments{
\item{path}{the path to the edited Insected and Curated file, for example,
"InspectAndCurate-Edited.csv"}
}
\description{
break the curated and edited file into different parts for later use and inspection
}
