% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_report_and_start_order.R
\name{join_report_and_start_order}
\alias{join_report_and_start_order}
\title{Full join music_report and start_order}
\usage{
join_report_and_start_order(MR, SO)
}
\arguments{
\item{MR}{the tibble of the music report file from EMS}

\item{SO}{the tibble of the start orders obtained from `scrape_starting_orders()`.}
}
\value{
This returns a tibble of the joined tables.  The second column
is `hasPerfectMatch`, the first column is status, which is full of
empty strings.  It is intended that this file will be manually curated
and appropriate entries placed in the `Status` column.
}
\description{
This does a full join on skater names and then for each event in the
music report that skater is doing, it checks for a perfect match in the
starting orders.  If it finds it, it will filter out all other matches
for that combination of skater name and music-report event (`Event`).
}
