% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_starting_orders.R
\name{scrape_starting_orders}
\alias{scrape_starting_orders}
\title{scrape PDFs of starting orders}
\usage{
scrape_starting_orders(Dir)
}
\arguments{
\item{Dir}{path to main directory.  Inside it you should find subfolders
which are named like `Friday`, `Saturday`, etc.  For example,
`Dir = "~/Downloads/2024 FCC Starting Orders"`.}
}
\description{
This puts information in a tibble and processes it somewhat, with
the hope that we can join it onto the music report at some point.
}
